feat(texas): Fully operational Texas merger command
Implement parsing step of Texas merger command and update to account for actual data structure.
