+++ .github/workflows/lint.yml
@@ -82,3 +82,6 @@ jobs:
           cl/search/docket_number_cleaner.py \
           cl/search/management/commands/clean_docket_number_raw.py \
           cl/scrapers/management/commands/back_scrape_dockets.py \
+          cl/corpus_importer/management/commands/import_texas_dockets.py \
+          cl/corpus_importer/management/utils.py \
+          cl/lib/decorators.py
